AP-1584

+++ i21Database/dbo/User Defined Types/VoucherDetailNonInvContract.sql
@@ -4,6 +4,6 @@
 	intContractDetailId INT NOT NULL,
 	intItemId			INT NOT NULL,
 	intAccountId		INT NOT NULL,
-	dblQtyReceived		INT NOT NULL DEFAULT 0,
-	dblCost				INT NOT NULL DEFAULT 0
+	dblQtyReceived		DECIMAL(18,6) NOT NULL DEFAULT 0,
+	dblCost				DECIMAL(18,6) NOT NULL DEFAULT 0
 )
